configuration

+++ scripts/start_ocms.sh
@@ -1,7 +1,15 @@
 #!/bin/bash
-LOGDIR="./logs"
-DBDIR="./db" 
+
+export OCMSPATH="/data/ocms2/dist"
+export AUTHPATH="/data/ocms2/dist"
+export OCMSLOGFILE="$OCMSPATH/logs/ocms.log"
+export AUTHLOGFILE="$AUTHPATH/logs/auth.log"
+export AUTHDBNAME="$AUTHPATH/db/auth.db"
+
 USER="pascal"
+LOGDIR="$AUTHPATH/logs"
+DBDIR="$AUTHPATH/db" 
+
 if [ ! -d "$LOGDIR" ]; then
 	echo "Create log directory"
     sudo mkdir -p $LOGDIR
@@ -10,18 +18,24 @@ if [ ! -d "$DBDIR" ]; then
 	echo "Create db directory"
     sudo mkdir -p $DBDIR    
 fi
+
 sudo chown -R $USER $LOGDIR
 sudo chown -R $USER $DBDIR
 echo "OCMS process started."
 if [ "$1" ==  "init" ]
  then
-     rm auth.log
- 	 echo "start auth init"
-     ./ocms init &
+ 	if [ -f "auth.log" ]
+	then
+	   rm auth.log
+	fi
+ 	echo "start auth init"
+    CMD="$OCMSPATH/ocms init &"
  else
- 	 echo "start auth"
-     ./ocms &
+ 	echo "start auth"
+    CMD="$OCMSPATH/ocms "
 fi
 
-read -rst 0.5
-tail -f ./logs/ocms.log
+eval "$CMD"
+
+#read -rst 0.5
+#tail -f $LOGDIR/ocms.log
